requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.2
aiohttp==3.8.3
requests==2.28.1
ipykernel==6.19.4
PyYAML==6.0